Backend/requirements.txt
Django
psycopg2-binary
djangorestframework
django-cors-headers
PyJWT
pandas==1.5.0
numpy
tabula-py
openpyxl
scikit-learn
tensorflow==2.13
opencv-python-headless==4.9.0.80
tifffile
click
joblib==1.3.2
matplotlib
patchify==0.2.3
pytest==8.1.1
h5py